+++ include/docker/docker_compose/zerotier-planet.sh
@@ -4,40 +4,40 @@ SRVS_PATH=${HOME}/Srvs
 mkdir ${SRVS_PATH}
 cd ${SRVS_PATH}
 
-SRV_NAME=nginx
+SRV_NAME=zerotier-planet
 mkdir ${SRV_NAME}
+
+docker image load < zerotier-planet.tar.gz
 
 docker run -d \
   --name ${SRV_NAME} \
   -e TZ=Asia/Shanghai \
-  nginx:latest
+  zerotier-planet
 
 echo "Waiting 3 secs for ${SRV_NAME} to start..."
 sleep 3
 
-mkdir -p ${SRV_NAME}/etc
-mkdir -p ${SRV_NAME}/usr/share
-docker container cp ${SRV_NAME}:/etc/nginx ${SRV_NAME}/etc
-docker container cp ${SRV_NAME}:/usr/share/nginx ${SRV_NAME}/usr/share
+mkdir -p ${SRV_NAME}/var/lib
+mkdir -p ${SRV_NAME}/app
+docker container cp ${SRV_NAME}:/var/lib/zerotier-one ${SRV_NAME}/var/lib
+docker container cp ${SRV_NAME}:/app ${SRV_NAME}
 docker container rm -f ${SRV_NAME}
 
 tee ${SRV_NAME}/compose.yaml <<EOF >>/dev/null
 version: '3'
 services:
   ${SRV_NAME}:
-    image: nginx:latest
+    image: zerotier-planet
     container_name: ${SRV_NAME}
     environment:
       - TZ=Asia/Shanghai
     ports:
-      - 80:80
-      - 443:443
+      - 4000:4000
+      - 9993:9993/udp
     volumes:
-      - ./usr/share/nginx/:/usr/share/nginx/
-      - ./etc/nginx/:/etc/nginx/
+      - ./var/lib/zerotier-one:/var/lib/zerotier-one
+      - ./app/backend/data:/app/backend/data
     restart: always
-    labels:
-      - "sh.acme.autoload.domain=zhw.link"
 EOF
 
 srv_include="  - ${SRV_NAME}/compose.yaml"